Merged revision(s) 368-369 from branches/Peach:
NOTE: IGNORE ANY CHANGES TO "IMAGES" LISTED IN THIS CHANGELOG, SVN WAS CONFLICTED

Completed - Issue 245: Page refresh after image upload creates duplicates
To combat this issue we noe operate a PRG (POST REDIRECT GET) pattern on the form processing stemming from the games and platform pages.

ADDITIONAL! A redirect is now in place on the login page to send the user back to the page that they clicked "Login" on.